refactor(data-availability): move checker into standalone crate
