+++ crates/common/data_availability/src/lib.rs
@@ -1,0 +1,5 @@
+mod checker;
+mod pending;
+
+pub use checker::DataAvailabilityChecker;
+pub use pending::{PendingAvailability, PendingBlock};
