+++ crates/common/fork_choice/beacon/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod constants;
-pub mod data_availability_checker;
 pub mod handlers;
 pub mod store;
+
+pub use ream_data_availability as data_availability;
